LayoutTests/http/tests/navigation/anchor-frames-cross-origin-expected.txt

CONSOLE MESSAGE: Fragment navigation not allowed with cross-origin frames.


--------
Frame: 'main'
--------
This prevents a cross-origin information leak sometimes know as framesniffing.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.body.offsetHeight > document.documentElement.clientHeight is true
PASS document.body.scrollTop == 0 is true
PASS document.body.scrollLeft == 0 is true
PASS successfullyParsed is true

TEST COMPLETE
This is an anchor point named "anchor1.

--------
Frame: 'footer'
--------
